+++ spec/lite/uxid/reversible/obfuscateid_spec.rb
@@ -2,17 +2,17 @@
 
 require "spec_helper"
 
-RSpec.describe Lite::Uxid::Reversible::Scatterid do
+RSpec.describe Lite::Uxid::Reversible::Obfuscateid do
   let(:prefix) { nil }
   let(:id) { 10 }
-  let(:scatterid) { "2056964183" }
-  let(:prefixed_scatterid) { "test_2056964183" }
+  let(:obfuscateid) { "2056964183" }
+  let(:prefixed_obfuscateid) { "test_2056964183" }
   let(:encoder) { described_class.new(id, prefix: prefix) }
-  let(:decoder) { described_class.new(scatterid, prefix: prefix) }
+  let(:decoder) { described_class.new(obfuscateid, prefix: prefix) }
 
   describe ".decode" do
     it "to be 10" do
-      expect(described_class.decode(scatterid)).to eq(id)
+      expect(described_class.decode(obfuscateid)).to eq(id)
       expect(decoder.decode).to eq(id)
     end
 
@@ -20,7 +20,7 @@ RSpec.describe Lite::Uxid::Reversible::Scatterid do
       let(:prefix) { "test_" }
 
       it "to be 10" do
-        expect(described_class.decode(scatterid, prefix: prefix)).to eq(id)
+        expect(described_class.decode(obfuscateid, prefix: prefix)).to eq(id)
         expect(decoder.decode).to eq(id)
       end
     end
@@ -28,8 +28,8 @@ RSpec.describe Lite::Uxid::Reversible::Scatterid do
 
   describe ".encode" do
     it "to be 2056964183" do
-      expect(described_class.encode(id)).to eq(scatterid)
-      expect(encoder.encode).to eq(scatterid)
+      expect(described_class.encode(id)).to eq(obfuscateid)
+      expect(encoder.encode).to eq(obfuscateid)
     end
 
     it "to be 10" do
@@ -41,8 +41,8 @@ RSpec.describe Lite::Uxid::Reversible::Scatterid do
       let(:prefix) { "test_" }
 
       it 'to be "test_2056964183"' do
-        expect(described_class.encode(id, prefix: prefix)).to eq(prefixed_scatterid)
-        expect(encoder.encode).to eq(prefixed_scatterid)
+        expect(described_class.encode(id, prefix: prefix)).to eq(prefixed_obfuscateid)
+        expect(encoder.encode).to eq(prefixed_obfuscateid)
       end
     end
   end