+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Reorganize generators into better namespaces
 - Fixed issue where nanoid would return without the prefix
+- Moved charsets to constants
 
 ## [1.5.2] - 2024-09-20
 ### Changed