+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.1] - 2024-09-23
+### Changed
+- Renamed `Scatterid` to `Obfuscateid`
+
 ## [2.0.0] - 2024-09-23
 ### Added
 - Scatterid reversible lib